This is very large update that should be considered unstable and shouldn't be used in a real project environment.  It is part of a larger feature set that will be submitted soon.
Changes:

1. Support for user configurable objects has been added, it is implemented using the UserConfigurableObject
   and UserConfigurableObjectField.  These define objects that are user configurable and the fields the
   object(s) can have.  They are backed by definition objects, UserConfigurableObjectType and
   UserConfigurableObjectTypeField.

2. Assets have been made user configurable as have chapters.  A LegacyUserConfigurableObject provides the
   glue between the existing "legacy" fields such as name and description and the asset.

3. Assets (Location, ResearchItem etc) have had their constructors made protected and the asset types should
   be created via a parent Project object.  The Project object contains the user configurable object type
   and type fields that the object needs for it's fields.  A single set of types and fields are used so that
   management of the fields is easier.

4. A number of sub-types of fields have been added to provide such for different data types such as Web pages,
   Selects, Dates, Multi-line text.  Some of the implementing classes will have their name changes before the
   full release so beware of what you use.

5. Basic, non-complete, support for adding fields to object types has been added however back-end support for
   saving those fields is not currently present.

6. The namedobject table has been modified to add support for a user object type so that all named objects
   could, in theory, support user configurable fields.  The object views (see data/schema/views/*.v) have been
   extended to expose the new field added to namedobject.  A side-effect of how the object manager works has
   required that the editor and project info schemas have also been updated to support the user object type key
   (this is a side-effect of how the tables are setup and re-use of the ObjectManager class for handling of
   NamedObjects).

+++ src/com/quollwriter/db/BookDataHandler.java
@@ -120,7 +120,7 @@ public class BookDataHandler implements DataHandler<Book, Project>
     }
 
     @Override
-    public Book getObjectByKey (int        key,
+    public Book getObjectByKey (long       key,
                                 Project    proj,
                                 Connection conn,
                                 boolean    loadChildObjects)
